subscriber/requirements.txt
pika==1.2.1
mysql-connector-python==8.0.32
